Intuition behind the problem
all algorithms I know for this problem follow the same pattern: let us start from the end and try to go in opposite direction. Imagine, that we have target = aabcaca, stamp = abca. Then we can go back and create a????ca string, then we can go to a?????? and finally to ???????. So, let us explain algorithm step by step:

    Let m and n be length of stamp and target.
    Let masks be binary masks for every window, when we put stamp on target. There will be n - m + 1 possible places for window. Window have 0 on place if stamp and target have the same element in given place. That is if bitmask equal to 0, we have full occurrence of stamp inside target.
    Now, we create these masks and also create set visited of indexes (starts of windows) and queue with candidates.
    We always keep in our queue candidates, such that they fully coincide with part of target (if we have ? symbol, it can be chosen as any symbol). We extract new candidate from left of our queue and add it to answer.
    Now, we want to understand, which elements we need to update, that is put ? to them. We have target_mask variable for this.
    Let us iterate over each element we need to update (for example in a????ca -> a?????? we need to update last two elements) and when we mark element as ?, we need to update all bits in masks[j]. Also we update target_mask.
    Let us found masks which are equal to 0 and add them to our queue if they are not visited yet.
    Finally, if target_mask is equal to zero, we return reversed answer and empty list in opposite case.

Complexity

Time complexity for first step is O((n-m)* n): first, we need to create masks, then for each of n elements, which we mark with ?, we need to spend O(m) iterations to do it: they will change this number of windows. Also to_update can have potential size m, but we will put every of n elements to and out of queue only once. Finally, part, where we are looking for zero masks have also O(m) complexity and repeated no more than m times. Final time complexity is O(n^2). Space complexity is O(n(n-m)).

Remark

Note, that we use bitmask with big size (more than 32 or 64) here and python allow us to do it. Moreover, when we update masks, at the moment we use loops, if we replace it with direct update it potentially will be much faster, I will try to do it later.